requirements.txt
pbr!=2.1.0,>=2.0.0 # Apache-2.0
requests>2.20.0 # Apache-2.0
Pillow>=6.0.0
webcolors>=1.9.1 # BSD 3-Clause
